refactor: extract DashboardService god class into focused query classes
Break the 816-line DashboardService into a thin facade (145 lines) plus
4 focused query classes, following the TelegramService refactor pattern
from PR #262:

- QueueDashboardQueries (90 lines) — queue detail, pending items
- MediaDashboardQueries (290 lines) — library, stats, categories, drift, content health
- HistoryDashboardQueries (320 lines) — analytics, schedule, approvals, team performance
- InstanceDashboardQueries (55 lines) — user instances, membership lookups

DashboardService keeps repos/init, resolve_chat_settings_id, thin
delegates preserving the public API, and get_service_health_stats
(system-level, not tenant-scoped).

Closes #253
